Added support for asdf 0.11.x

+++ lib/req.sh
@@ -138,8 +138,8 @@ _req1_with_asdf_inner_on_new_line()
 _req1_with_asdf_inner()
 {
     local program="$1"
-    local versionPolicy="$2"
-    local package="$3"
+    local versionPolicy="${2:-}"
+    local package="${3:-}"
     pluginName="$(_derive_asdf_plugin_name "$program" "$package")"
     emit_log "plugin $(ab "$pluginName"), "
     if ! _asdf_has_plugin "$pluginName"; then
@@ -198,9 +198,9 @@ _is_shim()
 
 _req1_with_asdf()
 {
-    local program="$1"
-    local versionPolicy="$2"
-    local package="$3"
+    local program="${1:-}"
+    local versionPolicy="${2:-}"
+    local package="${3:-}"
     if [ "$versionPolicy" = "$_VERSION_NO_CHECK" ] || [ "$versionPolicy" = "$_VERSION_ANY" ]; then
         emit_log "Checking $(ab "$program") ... "
         if _is_shim "$program"; then
@@ -271,7 +271,7 @@ _req1()
 _log_if_verbose()
 {
     local message=("$@")
-    if [ -n "$_REQ_VERBOSE" ]; then
+    if [ -n "${_REQ_VERBOSE:-}" ]; then
         log "${message[@]}"
     fi
 }
@@ -330,13 +330,13 @@ _compatible_packages()
 
 _req()
 {
-    if [[ $REQ_CHECKED = 1 ]]; then
+    if [[ "${REQ_CHECKED:-}" = 1 ]]; then
         exit_err "pre-boot script sanity checks already done, please define all requirements (i.e. all $(b req), $(b req_no_ver), and $(b req_ver) calls) before calling $(b req_check)"
     fi
     programName="$1"
     versionPolicy="$2"
-    package="$3"
-    if [[ $_REQ_INCLUDED = *" $programName:"* ]]; then
+    package="${3:-}"
+    if [[ "${_REQ_INCLUDED:-}" = *" $programName:"* ]]; then
         local includedPart=${_REQ_INCLUDED#*" $programName:"}
         local includedValue=${includedPart%%" "*}
         if [[ $includedValue != "$versionPolicy:$package" ]]; then
@@ -358,7 +358,7 @@ _req()
             _log_if_verbose "Found req for program $(ab "$programName"), versionPolicy=$(ab "$versionPolicy"), package=$(ab "$package") (already present)"
         fi
     else
-        _REQ_INCLUDED="$_REQ_INCLUDED $programName:$versionPolicy:$package"
+        _REQ_INCLUDED="${_REQ_INCLUDED:-} $programName:$versionPolicy:$package"
         _log_if_verbose "Found req for program $(ab "$programName"), versionPolicy=$(ab "$versionPolicy"), package=$(ab "$package")"
     fi
 }
@@ -398,7 +398,7 @@ req_ver()
     # - with asdf: try to install its latest matching version using asdf, otherwise fail
     local program="$1"
     local versionSpec="${2:-$_VERSION_ANY_VIA_ASDF_IF_AVAILABLE}"
-    local package="${3}"
+    local package="${3:-}"
     _req "$program" "$versionSpec" "$package"
 }
 
@@ -421,8 +421,13 @@ req_check()
     REQ_CHECKED=1
 
     log "Performing pre-boot script sanity checks [$(_describe_asdf_status)] ..."
-    tempVersions=()
     local programNameMarker="XXXXXXX"
+    local _temp_cached_asdf_plugin_list_file=""
+    if has_asdf; then
+        _temp_cached_asdf_plugin_list_file="$(mktemp -t "_asdf_plugin_list_cache_${programNameMarker}")"
+        _asdf_all_installed_plugins >"$_temp_cached_asdf_plugin_list_file" &
+    fi
+    tempVersions=()
     for entry in $_REQ_INCLUDED; do
         local program=${entry%%:*}
         local secondPart=${entry#*:}
@@ -430,25 +435,25 @@ req_check()
         local package=${secondPart##*:}
         local tempVersion
         tempVersion="$(mktemp -t "${program}${programNameMarker}")"
-        _req1 "$program" "$versionPolicy" "$package" 3>"$tempVersion" &
+        _cached_asdf_plugin_list_file="${_temp_cached_asdf_plugin_list_file:-}" _req1 "$program" "$versionPolicy" "$package" 3>"$tempVersion" &
         tempVersions+=("$tempVersion")
     done
     cleanup_temporary_version_files()
     {
         for tempVersion in "${tempVersions[@]}"; do
-            if [ -f "$tempVersion" ]; then
-                \rm "$tempVersion"
-            fi
+            [ -f "$tempVersion" ] && \rm "$tempVersion"
         done
+        [ -n "${_temp_cached_asdf_plugin_list_file:-}" ] && [ -f "$_temp_cached_asdf_plugin_list_file" ] && \rm "$_temp_cached_asdf_plugin_list_file"
     }
     trap cleanup_temporary_version_files RETURN
     wait
+    local failedRequirements=()
     for tempVersion in "${tempVersions[@]}"; do
         local simpleFileName="${tempVersion##*/}"
         local program="${simpleFileName%%"${programNameMarker}"*}"
         while IFS= read -r line; do
             if [ "$line" == "$_reqAbortMarker" ]; then
-                whine "Cowardly refusing to execute this script without the required program \"$program\". Have a nice day!"
+                failedRequirements+=("\"$program\"")
             fi
             local var
             var="$(echo "$line" | cut -d"=" -f1)"
@@ -458,5 +463,9 @@ req_check()
         done <"$tempVersion"
         rm "$tempVersion"
     done
-    log "$(green "Script sanity checks completed successfully, current script $(ab "$(tildify "$0")") can start.")"
+    if ((${#failedRequirements[@]} > 0)); then
+        whine "Cowardly refusing to execute this script without the required programs ${failedRequirements[*]}. Have a nice day!"
+    else
+        log "$(green "Script sanity checks completed successfully, current script $(ab "$(tildify "$0")") can start.")"
+    fi
 }
